pyproject.toml
[tool.poetry]
name = "strike-api"
version = "0.2.0"
description = "A python client for the strike api"
authors = ["Thomas Cross <tom.bz2@gmail.com>"]
readme = "README.md"
packages = [{include = "strike_api"}]

[tool.poetry.urls]
"Home Page" = "https://github.com/chmoder/strike-api"
"Bug Tracker" = "https://github.com/chmoder/strike-api/issues"
"Documentation" = "https://strike-api.readthedocs.io/en/latest/"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.1"
pydantic = "^1.10.2"
pyhumps = "^3.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
bandit = "^1.7.4"
sphinx = "^5.3.0"
black = "^22.10.0"
sphinx-rtd-theme = "^1.0.0"
pytest-recording = "^0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
